pyproject.toml
[project]
name = "weather-comparison"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.13",
    "coveralls>=4.0.1",
    "dash>=2.18.2",
    "pandas>=2.2.3",
    "polars>=1.22.0",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=19.0.1",
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
    "python-socketio>=5.12.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.37",
    "toml>=0.10.2",
]
